requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
supabase==2.0.0
sqlalchemy==2.0.23
gspread==5.12.0
oauth2client==4.1.3
loguru==0.7.2